requirements.txt
# Django and third-party apps
django==2.1.7
django-xff==1.3.0

# Web server
gunicorn==19.9.0

# Other utils
environs==4.1.2
mysqlclient==1.3.13
structlog==17.2.0
colorama==0.3.7
PyYAML==3.12
ipython==7.1.0

# Deployment
bumpversion==0.5.3

# Unpinned
raven

# Testing/development
pydevd-pycharm~=191.6605.12
django-debug-toolbar==1.10.1
django-debug-toolbar-template-profiler
django-extensions==2.1.4
django-queryinspect==1.0.0
factory_boy==2.8.1
loremipsum==1.0.5
mock==2.0.0
nose==1.3.7
testfixtures==4.10.0
autopep8
flake8
pycodestyle
